[ENH] move MultiEmbedding into layers/_embeddings
Relocate the shared embedding module out of models/nn and keep a
backward-compatible re-export so existing imports continue to work.
