+++ pytorch_forecasting/layers/_embeddings/__init__.py
@@ -6,6 +6,10 @@ from pytorch_forecasting.layers._embeddings._data_embedding import (
     DataEmbedding_inverted,
 )
 from pytorch_forecasting.layers._embeddings._en_embedding import EnEmbedding
+from pytorch_forecasting.layers._embeddings._multi_embedding import (
+    MultiEmbedding,
+    TimeDistributedEmbeddingBag,
+)
 from pytorch_forecasting.layers._embeddings._patch_embedding import PatchEmbedding
 from pytorch_forecasting.layers._embeddings._positional_embedding import (
     PositionalEmbedding,
@@ -15,7 +19,9 @@ from pytorch_forecasting.layers._embeddings._sub_nn import embedding_cat_variabl
 __all__ = [
     "DataEmbedding_inverted",
     "EnEmbedding",
+    "MultiEmbedding",
     "PositionalEmbedding",
+    "TimeDistributedEmbeddingBag",
     "embedding_cat_variables",
     "PatchEmbedding",
 ]
