+++ pytorch_forecasting/models/tide/_tide.py
@@ -13,7 +13,7 @@ from pytorch_forecasting.data import TimeSeriesDataSet
 from pytorch_forecasting.data.encoders import NaNLabelEncoder
 from pytorch_forecasting.metrics import MAE, MAPE, MASE, RMSE, SMAPE
 from pytorch_forecasting.models.base import BaseModelWithCovariates
-from pytorch_forecasting.models.nn.embeddings import MultiEmbedding
+from pytorch_forecasting.layers._embeddings import MultiEmbedding
 from pytorch_forecasting.models.tide.sub_modules import _TideModule
